pyproject.toml
[project]
name = "ocr-mcp"
version = "0.1.0"
description = "MCP server for OCR with multiple backends (Marker, DeepSeek, Mistral)"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "mcp>=0.9.0",
    "marker-pdf>=0.2.0",
    "httpx>=0.27.0",
    "python-dotenv>=1.0.0",
    "pillow>=10.0.0",
    "pypdf>=3.0.0",
    "pydantic>=2.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project.scripts]
ocr-mcp = "ocr_mcp.server:main"

[tool.uv]
dev-dependencies = []